+++ src/include/mem.h
@@ -1,8 +1,9 @@
 #ifndef MEM_H
 #define MEM_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-void memcpy(void *src, void *dest, uint64_t bytes);
+void memcpy(void *src, void *dest, size_t bytes);
 
 #endif
