+++ src/libc/stdio.c
@@ -2,6 +2,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 #include <ctype.h>
 #include <drivers/screen.h>
 #include <drivers/serial.h>
@@ -24,7 +25,7 @@ void vprintf(int device, const char* format, va_list arg)
     uint64_t l_val;
     char s_val[20];
 
-    for (int i = 0; i < strlen(format); i++) {
+    for (size_t i = 0; i < strlen(format); i++) {
         char c = format[i];
 
         if (c == '%') {
@@ -49,7 +50,7 @@ void vprintf(int device, const char* format, va_list arg)
                     i_val = va_arg(arg, int32_t);
                     itoa(i_val, s_val, 10);
 
-                    while ((leftpad - strlen(s_val)) > 0) {
+                    while ((leftpad - (uint8_t) strlen(s_val)) > 0) {
                         putchar(device, '0');
                         leftpad--;
                     }
