+++ src/libc/string.c
@@ -1,8 +1,8 @@
 #include <string.h>
 
-int strlen(const char* s)
+size_t strlen(const char* s)
 {
-    int ret = 0;
+    size_t ret = 0;
     while (s[ret] != 0) {
         ret++;
     }
@@ -13,7 +13,7 @@ int strlen(const char* s)
 void reverse(char* s)
 {
     char c;
-    for (int i = 0, j = strlen(s) - 1; i < j; i++, j--) {
+    for (size_t i = 0, j = strlen(s) - 1; i < j; i++, j--) {
         c = s[i];
         s[i] = s[j];
         s[j] = c;
